refactor: move security headers middleware to builder.go per review feedback
Move securityHeadersMiddleware to internal/app/builder.go and add it to
the default middleware chain in buildHTTPServer, alongside the other
default middlewares (RequestID, Recoverer, Timeout, LoggingMiddleware).

Revert the server.go changes — NewServer remains unchanged and the
security headers are applied through the builder's default middleware
chain as intended.

Move the test to internal/app/builder_test.go accordingly.
